src/requirements.txt
django
django-cors-headers
django-wysiwyg-redactor
boto
psycopg2
requests
pillow
